+++ app/src/main/java/com/example/mge_menstracker/DayViewContainer.kt
@@ -1,0 +1,9 @@
+package com.example.mge_menstracker
+
+import android.view.View
+import android.widget.TextView
+import com.kizitonwose.calendarview.ui.ViewContainer
+
+class DayViewContainer(view: View) : ViewContainer(view) {
+    val textView = view.findViewById<TextView>(R.id.calendarDayText)
+}
