+++ app/src/main/res/layout/activity_newcycle.xml
